4. Модуляция с исправлениями
