+++ js/data.js
@@ -1,3 +1,5 @@
+import {getRandomDecimalInclusive, getRandomArrayElement, getRandomArrayIndicators} from './utils.js'
+
 const avatar = [
   'img/avatars/user01.png',
   'img/avatars/user02.png',
@@ -66,4 +68,33 @@ const photos = [
   'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
 ]
 
-export {avatar, title, price, type, rooms, guests, checkin, checkout, features, description, photos};
+const apartmentDeals = () => {
+  return [
+    {
+      author : {
+        avatar : getRandomArrayElement(avatar),
+      },
+
+      offer : {
+        title : title,
+        address : [location.x, location.y],
+        price : getRandomArrayElement(price),
+        type : getRandomArrayElement(type),
+        rooms : getRandomArrayElement(rooms),
+        guests : getRandomArrayElement(guests),
+        checkin : getRandomArrayElement(checkin),
+        checkout : getRandomArrayElement(checkout),
+        features : getRandomArrayIndicators(features),
+        description : description,
+        photos : getRandomArrayIndicators(photos),
+      },
+
+      location : {
+        x : getRandomDecimalInclusive(35.65000, 35.70000, 5),
+        y : getRandomDecimalInclusive(139.70000, 139.80000, 5),
+      },
+    },
+  ];
+};
+
+export {apartmentDeals};
